[FIX] sale_margin_delivered: Uses the right field to get price reduce
'price_reduce' is deprecated and removed into the next version. Uses the 'price_reduce_taxexcl' field instead.

As an additional benefit this change ensures the compatibility with the 'sale_triple_discount' addon. Indeed, when 'sale_triple_discount' is installed, the dicount field is not used as an aggregation of all the applied discount. It's only use to store the first discount applied. Therefore, the field is not properly computed since it doesn't include the second and third discound. On the other side 'price_reduce_taxexcl' is computed from the 'price_subtotal' which includes all the discounts